[GStreamer] Remove useless workaround https://bugs.webkit.org/show_bug.cgi?id=186921
Reviewed by Xabier Rodriguez-Calvar.

In bug 67407 a workaround was added for GStreamer 0.10. With 1.x
the media/video-reverse-play-duration.html test passes without any
workaround needed. The other test mentioned in that bug was
removed, it seems.

* platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.cpp:
(WebCore::MediaPlayerPrivateGStreamer::playbackPosition const):
Removed early return, position queries shouldn't be avoided on EOS
because the pipeline is in READY state, not NULL.
(WebCore::MediaPlayerPrivateGStreamer::currentMediaTime const):

Identifier: 202701@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233637 268f45cc-cd09-0410-ab3c-d52691b4dbfc