[Flatpak] Feature options to build-webkit and --makeargs are not passed inside the sandbox https://bugs.webkit.org/show_bug.cgi?id=187470
Patch by Thibault Saunier <tsaunier@igalia.com> on 2018-07-10
Reviewed by Philippe Normand.

Never force building, the user is responsible to make that decision and the workflow is
now based on external scripts so it doesn't make sense anymore. Doing that leads
to ignoring the options the user passed in.

* flatpak/flatpakutils.py:
(WebkitFlatpak.setup_dev_env):

Identifier: 202743@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233681 268f45cc-cd09-0410-ab3c-d52691b4dbfc